chore(chat): allow fractional INDEXER_INTERVAL_HOURS for sub-hour polling (#885)
Widen indexer_interval_hours from int to float so the indexer can poll more
often than once an hour (e.g. 0.25 = every 15 min) — letting freshly published
library.db / attributions go live in minutes instead of up to 6 hours. The
loop still floors the effective interval at 60s. Prod value is set via
/opt/lectorium/.env (INDEXER_INTERVAL_HOURS); .env.example documents the
fractional option.

+++ modules/services/chat/app/src/lectorium_chat/config.py
@@ -141,7 +141,9 @@ class Settings(BaseSettings):
 
     # ── Indexer ─────────────────────────────────────────────────────────
     catalog_dir: Path = Path("/var/lib/chat")
-    indexer_interval_hours: int = 6
+    # Fractional values are allowed (e.g. 0.25 = every 15 min). The loop
+    # floors the effective interval at 60s regardless (see indexer/run.py).
+    indexer_interval_hours: float = 6
     indexer_langs: str = "ru,en"
     indexer_bootstrap_on_start: bool = True
